* Don't use bullet sign in addon name * Minimize entry point Python file for better performances  (as suggested by enen92)
